H3DViewer - RebuildDisplayLists were not done often enough. Also fixed so that all the settings in the setting.ini file is not reloaded for each new opened file. This fixes some annoying features.
Sequencers - Removed two routes to the value_changed field in order to not trigger updates of this field when there should be none.

Corrected doxygen in GeneratedCubeMapTexture.h and removed unused variables and includes from some files.

Fixed issues with near and far clip-plane in H3DWindowNode when putting the viewpoint inside a transform that scales. Also updated H3DHapticsDevice so that followViewpoint works for these cases.
Also moved pushMatrix and popMatrix pairs in H3DWindowNode.cpp  inside an if case (that was between the pairs) since nothing was done unless the if case was evaluated to true.

The avatarSize field is now scaled with scaling of viewpoint as per X3D specification.

git-svn-id: https://www.h3d.org:8090/H3DAPI/trunk/H3DAPI@1129 610dfd11-2700-0410-9b68-9ae28eac5983