Implemented vertex buffer object support in ElevationGrid and SuperShape. Made some updates to the other vbo-parts (just named an internal field, for future documentation).
Extrusion - Rearranged and made it to code standard. It is now more consistent with other nodes. Remade it to be able to make it easy to implement vbo-support but in the end realized that it would only work in a very very special case without a lot of extra data stored.

git-svn-id: https://www.h3d.org:8090/H3DAPI/trunk/H3DAPI@1611 610dfd11-2700-0410-9b68-9ae28eac5983