DirectShowDecoder.cpp - Fixed bug with getPosition (report 0000286) not returning correct value. Also fixed memory leak.
Node, Sphere, Text, X3DGeometryNode - Made optimizations since bug report 0000291 pointed out that a lot of extra call was made. One can not make as much optimizations as is done when hiding part of the scene inside collision/ToggleGroups but one can at least avoid to do collision with the boundTree in those cases when there is no pointingdevicesensor.

H3DNavigation was checking pointingdevicesensors for some reason, that is not needed at all.

CylinderSensor, SphereSensor, PlaneSensor - Made sure that Spheresensor behaves like the rest of them when it comes to offset and fixed issue with CylinderSensor offset value not being used.

git-svn-id: https://www.h3d.org:8090/H3DAPI/trunk/H3DAPI@1957 610dfd11-2700-0410-9b68-9ae28eac5983