WxFrame.cpp, SpiderMonkeyTypes, Script, SpiderMonkeySAI - Fixed memory leaks.
GLUTWindow.h - Documentation update.

ProtoDeclaration, X3DSAX2Handlers - Made sure that protos are aware of other protos that are in the same level of the hierarchy or above. As far as I can understand of the X3D specification this is not invalid use. In X3DSAX2Handlers I also allowed ProtoInstance to have IS statements (and connect statements in the IS) since a ProtoBody can contain a ProtoInstance.

CheckExample.py - Just made printouts neater.

Sphere.cpp - Removed unneccesary warning.

git-svn-id: https://www.h3d.org:8090/H3DAPI/trunk/H3DAPI@2006 610dfd11-2700-0410-9b68-9ae28eac5983