WxFrame - Fixed some annoying constructor calls which gave "this, this" as input. Just changed it to take only one input.
The rest, Added a small fix to make H3DViewer go into fullscreen directly without rendering one black frame first. This does not work perfectly because of a bug in WxWidgets which makes it not possible to go back to windowed mode if ShowFullScreen is called on a top level window before Show.

This is not tested on linux yet... Might be that the correct call order works on linux.

git-svn-id: https://www.h3d.org:8090/H3DAPI/trunk/H3DAPI@2561 610dfd11-2700-0410-9b68-9ae28eac5983