H3DFunctionNode - documentation correction, these fields are not exposed at X3D/python level.
MField, SField - setValueFromVoidPtr now takes const void * as input instead of void *. Since it is a virtual function the old signature is still there but deprecated. Also made sure that Node * calls the correct version of these functions.

X3DTextureNode - Made sure that the build does not fail if OpenEXR does not exist.

git-svn-id: https://www.h3d.org:8090/H3DAPI/trunk/H3DAPI@2781 610dfd11-2700-0410-9b68-9ae28eac5983